Temporary "OrganizationWithDaoTest" cleaned up
